+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- `NEW` Changed published messages can be reverted to published state and archived
+- `NEW` Changes on published messages can be discarded
 - `NEW` Published messages can be archived (taken offline)
 - `CNG` Administrative areas are preloaded differently
 - `CNG` Login and other user related pages are centered on the screen